+++ app/src/main/java/com/damasahhre/hooftrim/database/dao/InjuryDao.java
@@ -19,7 +19,7 @@ public interface InjuryDao {
             "AND Report.leg_area_number == 0 " +
             "AND Report.reference_cause_new_limp " +
             "GROUP BY Report.cow_id, Report.visit_date ")
-    Integer felemons(Integer farmId, MyDate start, MyDate end);
+    Integer felemons(Long farmId, MyDate start, MyDate end);
 
     @Query("SELECT COUNT(Report.id) FROM Report,Cow,Farm " +
             "WHERE Report.cow_id == Cow.id " +
@@ -30,7 +30,7 @@ public interface InjuryDao {
             "AND Report.leg_area_number == 10 " +
             "AND Report.reference_cause_new_limp " +
             "GROUP BY Report.cow_id, Report.visit_date ")
-    Integer deramatit(Integer farmId, MyDate start, MyDate end);
+    Integer deramatit(Long farmId, MyDate start, MyDate end);
 
     @Query("SELECT Report.cow_id AS cowId, Report.visit_date AS date, Report.finger_number AS fingerNumber " +
             "FROM Report,Cow,Farm " +
@@ -42,7 +42,7 @@ public interface InjuryDao {
             "AND Report.leg_area_number == 4 " +
             "AND Report.reference_cause_new_limp " +
             "GROUP BY Report.cow_id, Report.visit_date ")
-    List<InjureyReport> woundHoofBottom(Integer farmId, MyDate start, MyDate end);
+    List<InjureyReport> woundHoofBottom(Long farmId, MyDate start, MyDate end);
 
     @Query("SELECT Report.cow_id AS cowId, Report.visit_date AS date, Report.finger_number AS fingerNumber" +
             " FROM Report,Cow,Farm " +
@@ -54,7 +54,7 @@ public interface InjuryDao {
             "AND (Report.leg_area_number == 2 OR Report.leg_area_number == 3) " +
             "AND Report.reference_cause_new_limp " +
             "GROUP BY Report.finger_number,Report.cow_id, Report.visit_date ")
-    List<InjureyReport> whiteLineWound(Integer farmId, MyDate start, MyDate end);
+    List<InjureyReport> whiteLineWound(Long farmId, MyDate start, MyDate end);
 
     @Query("SELECT Report.cow_id AS cowId, Report.visit_date AS date, Report.finger_number AS fingerNumber" +
             " FROM Report,Cow,Farm " +
@@ -66,7 +66,7 @@ public interface InjuryDao {
             "AND (Report.leg_area_number == 5 OR Report.leg_area_number == 1)" +
             "AND Report.reference_cause_new_limp " +
             "GROUP BY Report.finger_number, Report.cow_id, Report.visit_date ")
-    List<InjureyReport> pangeWound(Integer farmId, MyDate start, MyDate end);
+    List<InjureyReport> pangeWound(Long farmId, MyDate start, MyDate end);
 
     @Query("SELECT Report.cow_id AS cowId, Report.visit_date AS date, Report.finger_number AS fingerNumber" +
             " FROM Report,Cow,Farm " +
@@ -78,7 +78,7 @@ public interface InjuryDao {
             "AND Report.leg_area_number == 6 " +
             "AND Report.reference_cause_new_limp " +
             "GROUP BY Report.finger_number, Report.cow_id, Report.visit_date ")
-    List<InjureyReport> pashneWound(Integer farmId, MyDate start, MyDate end);
+    List<InjureyReport> pashneWound(Long farmId, MyDate start, MyDate end);
 
     @Query("SELECT COUNT(Report.id) FROM Report,Cow,Farm " +
             "WHERE Report.cow_id == Cow.id " +
@@ -90,7 +90,7 @@ public interface InjuryDao {
             "OR Report.leg_area_number == 11 OR Report.leg_area_number == 12) " +
             "AND Report.reference_cause_new_limp " +
             "GROUP BY Report.finger_number, Report.cow_id, Report.visit_date ")
-    Integer wallWound(Integer farmId, MyDate start, MyDate end);
+    Integer wallWound(Long farmId, MyDate start, MyDate end);
 
     @Query("SELECT COUNT(Report.id) FROM Report,Cow,Farm " +
             "WHERE Report.cow_id == Cow.id " +
@@ -101,7 +101,7 @@ public interface InjuryDao {
             "AND Report.leg_area_number == 9 " +
             "AND Report.reference_cause_new_limp " +
             "GROUP BY Report.cow_id, Report.visit_date ")
-    Integer reigenNine(Integer farmId, MyDate start, MyDate end);
+    Integer reigenNine(Long farmId, MyDate start, MyDate end);
 
     @Query("SELECT Report.id FROM Report,Cow,Farm " +
             "WHERE Report.cow_id == Cow.id " +
@@ -109,7 +109,7 @@ public interface InjuryDao {
             "AND Farm.id == :farmId " +
             "AND Report.visit_date >= :date  AND Report.visit_date <= :endDate  " +
             "GROUP BY Report.cow_id")
-    List<Integer> box(Integer farmId, MyDate date, MyDate endDate);
+    List<Integer> box(Long farmId, MyDate date, MyDate endDate);
 
     @Query("SELECT Report.id FROM Report,Cow,Farm " +
             "WHERE Report.cow_id == Cow.id " +
@@ -118,7 +118,7 @@ public interface InjuryDao {
             "AND Report.visit_date >= :date  AND Report.visit_date <= :endDate  " +
             "AND Report.reference_cause_limp_visit " +
             "GROUP BY Report.cow_id")
-    List<Integer> visit(Integer farmId, MyDate date, MyDate endDate);
+    List<Integer> visit(Long farmId, MyDate date, MyDate endDate);
 
     @Query("SELECT Report.id FROM Report,Cow,Farm " +
             "WHERE Report.cow_id == Cow.id " +
@@ -127,7 +127,7 @@ public interface InjuryDao {
             "AND Report.visit_date >= :date  AND Report.visit_date <= :endDate  " +
             "AND Report.reference_cause_new_limp " +
             "GROUP BY Report.cow_id")
-    List<Integer> newLimp(Integer farmId, MyDate date, MyDate endDate);
+    List<Integer> newLimp(Long farmId, MyDate date, MyDate endDate);
 
     @Query("SELECT Report.id FROM Report,Cow,Farm " +
             "WHERE Report.cow_id == Cow.id " +
@@ -136,7 +136,7 @@ public interface InjuryDao {
             "AND Report.visit_date >= :date  AND Report.visit_date <= :endDate  " +
             "AND Report.reference_cause_hundred_days " +
             "GROUP BY Report.cow_id")
-    List<Integer> sadRoze(Integer farmId, MyDate date, MyDate endDate);
+    List<Integer> sadRoze(Long farmId, MyDate date, MyDate endDate);
 
     @Query("SELECT Report.id FROM Report,Cow,Farm " +
             "WHERE Report.cow_id == Cow.id " +
@@ -145,7 +145,7 @@ public interface InjuryDao {
             "AND Report.visit_date >= :date  AND Report.visit_date <= :endDate  " +
             "AND Report.reference_cause_dryness " +
             "GROUP BY Report.cow_id")
-    List<Integer> dryness(Integer farmId, MyDate date, MyDate endDate);
+    List<Integer> dryness(Long farmId, MyDate date, MyDate endDate);
 
     @Query("SELECT Report.id FROM Report,Cow,Farm " +
             "WHERE Report.cow_id == Cow.id " +
@@ -154,7 +154,7 @@ public interface InjuryDao {
             "AND Report.visit_date >= :date  AND Report.visit_date <= :endDate  " +
             "AND Report.reference_cause_lagged " +
             "GROUP BY Report.cow_id")
-    List<Integer> delayed(Integer farmId, MyDate date, MyDate endDate);
+    List<Integer> delayed(Long farmId, MyDate date, MyDate endDate);
 
     @Query("SELECT Report.id FROM Report,Cow,Farm " +
             "WHERE Report.cow_id == Cow.id " +
@@ -163,7 +163,7 @@ public interface InjuryDao {
             "AND Report.visit_date >= :date  AND Report.visit_date <= :endDate  " +
             "AND Report.reference_cause_group_hoof_trim " +
             "GROUP BY Report.cow_id")
-    List<Integer> group(Integer farmId, MyDate date, MyDate endDate);
+    List<Integer> group(Long farmId, MyDate date, MyDate endDate);
 
     @Query("SELECT Report.id FROM Report,Cow,Farm " +
             "WHERE Report.cow_id == Cow.id " +
@@ -172,7 +172,7 @@ public interface InjuryDao {
             "AND Report.visit_date >= :date  AND Report.visit_date <= :endDate  " +
             "AND Report.reference_cause_high_score " +
             "GROUP BY Report.cow_id")
-    List<Integer> high(Integer farmId, MyDate date, MyDate endDate);
+    List<Integer> high(Long farmId, MyDate date, MyDate endDate);
 
     @Query("SELECT Report.id FROM Report,Cow,Farm " +
             "WHERE Report.cow_id == Cow.id " +
@@ -181,7 +181,7 @@ public interface InjuryDao {
             "AND Report.visit_date >= :date  AND Report.visit_date <= :endDate  " +
             "AND Report.reference_cause_referential " +
             "GROUP BY Report.cow_id")
-    List<Integer> refrence(Integer farmId, MyDate date, MyDate endDate);
+    List<Integer> refrence(Long farmId, MyDate date, MyDate endDate);
 
     @Query("SELECT Report.id FROM Report,Cow,Farm " +
             "WHERE Report.cow_id == Cow.id " +
@@ -190,7 +190,7 @@ public interface InjuryDao {
             "AND Report.visit_date >= :date  AND Report.visit_date <= :endDate  " +
             "AND Report.reference_cause_heifer " +
             "GROUP BY Report.cow_id")
-    List<Integer> heifer(Integer farmId, MyDate date, MyDate endDate);
+    List<Integer> heifer(Long farmId, MyDate date, MyDate endDate);
 
     @Query("SELECT Report.id FROM Report,Cow,Farm " +
             "WHERE Report.cow_id == Cow.id " +
@@ -199,7 +199,7 @@ public interface InjuryDao {
             "AND Report.visit_date >= :date  AND Report.visit_date <= :endDate  " +
             "AND Report.reference_cause_long_hoof " +
             "GROUP BY Report.cow_id")
-    List<Integer> longHoof(Integer farmId, MyDate date, MyDate endDate);
+    List<Integer> longHoof(Long farmId, MyDate date, MyDate endDate);
 
     @Query("SELECT Report.id FROM Report,Cow,Farm " +
             "WHERE Report.cow_id == Cow.id " +
@@ -208,7 +208,7 @@ public interface InjuryDao {
             "AND Report.visit_date >= :date  AND Report.visit_date <= :endDate  " +
             "AND Report.other_info_hoof_trim " +
             "GROUP BY Report.cow_id")
-    List<Integer> somChini(Integer farmId, MyDate date, MyDate endDate);
+    List<Integer> somChini(Long farmId, MyDate date, MyDate endDate);
 
     @Query("SELECT Report.id FROM Report,Cow,Farm " +
             "WHERE Report.cow_id == Cow.id " +
@@ -217,7 +217,7 @@ public interface InjuryDao {
             "AND Report.visit_date >= :date  AND Report.visit_date <= :endDate  " +
             "AND Report.other_info_boarding " +
             "GROUP BY Report.cow_id")
-    List<Integer> boarding(Integer farmId, MyDate date, MyDate endDate);
+    List<Integer> boarding(Long farmId, MyDate date, MyDate endDate);
 
 
     //
@@ -229,7 +229,7 @@ public interface InjuryDao {
             "AND Farm.id == :farmId " +
             "AND Report.visit_date >= :date  AND Report.visit_date <= :endDate  " +
             "AND Report.other_info_boarding ")
-    List<Integer> boardingFactorAll(Integer farmId, MyDate date, MyDate endDate);
+    List<Integer> boardingFactorAll(Long farmId, MyDate date, MyDate endDate);
 
     @Query("SELECT Report.id FROM Report,Cow,Farm " +
             "WHERE Report.cow_id == Cow.id " +
@@ -238,7 +238,7 @@ public interface InjuryDao {
             "AND Farm.id == :farmId " +
             "AND Report.visit_date >= :date  AND Report.visit_date <= :endDate  " +
             "AND Report.other_info_boarding ")
-    List<Integer> boardingFactor(Integer farmId, MyDate date, MyDate endDate, Integer cowNumber);
+    List<Integer> boardingFactor(Long farmId, MyDate date, MyDate endDate, Integer cowNumber);
 
     @Query("SELECT Report.id FROM Report,Cow,Farm " +
             "WHERE Report.cow_id == Cow.id " +
@@ -246,7 +246,7 @@ public interface InjuryDao {
             "AND Farm.id == :farmId " +
             "AND Report.visit_date >= :date  AND Report.visit_date <= :endDate  " +
             "AND Report.other_info_hoof_trim ")
-    List<Integer> hoofTrimFactorAll(Integer farmId, MyDate date, MyDate endDate);
+    List<Integer> hoofTrimFactorAll(Long farmId, MyDate date, MyDate endDate);
 
     @Query("SELECT Report.id FROM Report,Cow,Farm " +
             "WHERE Report.cow_id == Cow.id " +
@@ -255,7 +255,7 @@ public interface InjuryDao {
             "AND Farm.id == :farmId " +
             "AND Report.visit_date >= :date  AND Report.visit_date <= :endDate  " +
             "AND Report.other_info_hoof_trim ")
-    List<Integer> hoofTrimFactor(Integer farmId, MyDate date, MyDate endDate, Integer cowNumber);
+    List<Integer> hoofTrimFactor(Long farmId, MyDate date, MyDate endDate, Integer cowNumber);
 
     @Query("SELECT Report.id FROM Report,Cow,Farm " +
             "WHERE Report.cow_id == Cow.id " +
@@ -263,7 +263,7 @@ public interface InjuryDao {
             "AND Farm.id == :farmId " +
             "AND Report.visit_date >= :date  AND Report.visit_date <= :endDate  " +
             "AND Report.other_info_gel ")
-    List<Integer> gelFactorAll(Integer farmId, MyDate date, MyDate endDate);
+    List<Integer> gelFactorAll(Long farmId, MyDate date, MyDate endDate);
 
     @Query("SELECT Report.id FROM Report,Cow,Farm " +
             "WHERE Report.cow_id == Cow.id " +
@@ -272,7 +272,7 @@ public interface InjuryDao {
             "AND Farm.id == :farmId " +
             "AND Report.visit_date >= :date  AND Report.visit_date <= :endDate  " +
             "AND Report.other_info_gel ")
-    List<Integer> gelFactor(Integer farmId, MyDate date, MyDate endDate, Integer cowNumber);
+    List<Integer> gelFactor(Long farmId, MyDate date, MyDate endDate, Integer cowNumber);
 
     @Query("SELECT Report.id FROM Report,Cow,Farm " +
             "WHERE Report.cow_id == Cow.id " +
@@ -280,7 +280,7 @@ public interface InjuryDao {
             "AND Farm.id == :farmId " +
             "AND Report.visit_date >= :date  AND Report.visit_date <= :endDate  " +
             "AND Report.reference_cause_limp_visit ")
-    List<Integer> visitFactorAll(Integer farmId, MyDate date, MyDate endDate);
+    List<Integer> visitFactorAll(Long farmId, MyDate date, MyDate endDate);
 
     @Query("SELECT Report.id FROM Report,Cow,Farm " +
             "WHERE Report.cow_id == Cow.id " +
@@ -289,6 +289,6 @@ public interface InjuryDao {
             "AND Farm.id == :farmId " +
             "AND Report.visit_date >= :date  AND Report.visit_date <= :endDate  " +
             "AND Report.reference_cause_limp_visit ")
-    List<Integer> visitFactor(Integer farmId, MyDate date, MyDate endDate, Integer cowNumber);
+    List<Integer> visitFactor(Long farmId, MyDate date, MyDate endDate, Integer cowNumber);
 
 }
